requirements.txt
PyYAML==5.4.1
requests==2.25.1
streamlit==0.81.1

